MERGE PROJECTS INTO ONE BIG PROJECT
This hard merged all the projects for the GEAR device into one repo.
Most code is not well tested and documented, but an updated README file was
added. This would communicate the barebones necessities of the project and
also the need. It would overview the high level design we basically
implemented.